refactor: coderabbit suggestions
