--- packages/cli/src/onboarding/types.ts
@@ -1,1 +1,0 @@
-export type AvailiblePackages = '@c15t/nextjs' | '@c15t/react' | 'c15t';
